Refinement of the project structure
